--- a/packages/stylist-core/src/lib.rs
+++ b/packages/stylist-core/src/lib.rs
@@ -9,6 +9,7 @@
 mod error;
 pub use error::{Error, Result, ResultDisplay};
 pub mod ast;
+pub mod bow;
 
 #[cfg_attr(documenting, doc(cfg(feature = "parser")))]
 #[cfg(feature = "parser")]