--- a/packages/stylist-macros/src/literal/argument.rs
+++ b/packages/stylist-macros/src/literal/argument.rs
@@ -1,7 +1,7 @@
 use proc_macro2::{Ident, TokenStream};
 
-#[derive(Debug)]
-pub(crate) struct Argument {
+#[derive(Debug, Clone)]
+pub struct Argument {
     pub name: String,
     pub name_token: Ident,
     pub tokens: TokenStream,