--- a/packages/stylist-macros/src/output/context_recorder.rs
+++ b/packages/stylist-macros/src/output/context_recorder.rs
@@ -1,7 +1,7 @@
 //! This module implements a type abstractly tracking in what kind of expression context
 //! an item appears. This information is leverage to provide improved performance and
 //! static caching of parts of the generated output.
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
 enum AllowedUsage {
     // ```
     // let width = 500;
@@ -14,13 +14,10 @@ enum AllowedUsage {
     // //       ------------- everything is static, do wrap style in Lazy
     // ```
     Static,
-    // TODO: we can probably avoid a few allocations if we track which parts
+    // We can avoid a few allocations if we track which parts
     // of the ast can be constructed statically (with const methods), which is
     // even stronger than constructing it in the global context in a Lazy.
-    // Should you decide to implement this, keep in mind to change Self::MAX
-    // and adjust the generation of cow-vec tokens. Also check the usages of
-    // MaybeStatic::statick if they can be upgraded to Const.
-    // Const,
+    Const,
 }
 
 #[derive(Debug, Clone)]
@@ -31,7 +28,7 @@ pub struct ContextRecorder {
 impl Default for ContextRecorder {
     fn default() -> Self {
         Self {
-            usage: AllowedUsage::Static,
+            usage: AllowedUsage::Const,
         }
     }
 }
@@ -41,12 +38,24 @@ impl ContextRecorder {
         Self::default()
     }
 
+    pub fn uses_nested(&mut self, other: &Self) {
+        self.usage = self.usage.min(other.usage);
+    }
+
     // Record the usage of a dynamic expression
     pub fn uses_dynamic_argument(&mut self) {
-        self.usage = AllowedUsage::Dynamic;
+        self.usage = self.usage.min(AllowedUsage::Dynamic);
+    }
+    // Record the usage of an expression that is not allowed in const context
+    pub fn uses_static(&mut self) {
+        self.usage = self.usage.min(AllowedUsage::Static);
     }
 
     pub fn is_static(&self) -> bool {
-        self.usage == AllowedUsage::Static
+        self.usage >= AllowedUsage::Static
+    }
+
+    pub fn is_const(&self) -> bool {
+        self.usage >= AllowedUsage::Const
     }
 }