--- a/packages/stylist-macros/src/output/maybe_static.rs
+++ b/packages/stylist-macros/src/output/maybe_static.rs
@@ -1,5 +1,5 @@
 use super::{ContextRecorder, Reify};
-use proc_macro2::TokenStream;
+use proc_macro2::{Ident, Span, TokenStream};
 use quote::quote;
 
 pub trait IntoCowVecTokens: IntoIterator
@@ -10,7 +10,7 @@ where
     // as elements the values formed by the expressions in this stream.
     // Depending on the context in which the expression can be expanded,
     // uses either Cow::Owned or Cow::Borrowed (currently always Cow::Owned).
-    fn into_cow_vec_tokens(self, ctx: &mut ContextRecorder) -> TokenStream;
+    fn into_cow_vec_tokens(self, typ: TokenStream, ctx: &mut ContextRecorder) -> TokenStream;
 }
 
 impl<I> IntoCowVecTokens for I
@@ -18,23 +18,35 @@ where
     I: IntoIterator,
     I::Item: Reify,
 {
-    fn into_cow_vec_tokens(self, ctx: &mut ContextRecorder) -> TokenStream {
-        let contents: Vec<TokenStream> =
-            self.into_iter().map(|m| m.into_token_stream(ctx)).collect();
+    fn into_cow_vec_tokens(self, typ: TokenStream, ctx: &mut ContextRecorder) -> TokenStream {
+        let mut inner_ctx = ContextRecorder::new();
+        let contents: Vec<TokenStream> = self
+            .into_iter()
+            .map(|m| m.into_token_stream(&mut inner_ctx))
+            .collect();
+        ctx.uses_nested(&inner_ctx);
 
-        quote! {
-            ::std::vec![
-                #( #contents, )*
-            ].into()
+        if inner_ctx.is_const() {
+            let name = Ident::new("items", Span::mixed_site());
+            let content_len = contents.len();
+            quote! {
+                ::std::borrow::Cow::<[#typ]>::Borrowed (
+                    {
+                        const #name: [#typ; #content_len] = [
+                            #( #contents, )*
+                        ];
+                        &#name
+                    }
+                )
+            }
+        } else {
+            quote! {
+                ::std::borrow::Cow::<[#typ]>::Owned (
+                    ::std::vec![
+                        #( #contents, )*
+                    ]
+                )
+            }
         }
-
-        // In the future, if there's a need to collect sub contexts for optimisation:
-        // let tokens = TokenStream::new();
-        // let mut ctx = ContextRecorder::new();
-        // for i in self.into_iter() {
-        //     tokens.extend(i.into_token_stream(&mut ctx));
-        //     tokens.extend(quote! {, });
-        // }
-        // quote! { ::std::vec![#items] }
     }
 }