Changes the usages of `Vec` in favor of `const` arrays.